rewrite macro with global execution mask to support conditionals
I'm not super happy about this strategy, as it means we're giving up the
efficiency of only rewriting LCDs. But that was never quite correct—the
right thing to do would be to rewrite all assignments where the variable
assigned will make it out of the masked scope, and only those. But that
would require more static analysis, so I'll get to that eventually.

Should we also use the global mask for `for` loops, or is the current
behavior (relying on propagation of the mask associated with the
iteration variable) correct and faster? I'm leaning towards the latter.