add batch dimension indexing sentinel BD
We need this because a subtle distinction between NumPy/PyTorch's basic
slicing and advanced indexing behavior means that no single value for
the batch dimension index will get us the right behavior for
example-level programming, given that we're working with a user-visible
batch dimension.

The Matchbox equivalent of truly example-level `arr[0]` (basic slicing)
is `arr[:, 0]`, but the equivalent of `arr[[0]]` is `arr[[0], [0]]`.
With the sentinel they can be written as `arr[BD, 0]` and `arr[BD,
[0]]`.